+++ packages/@baldeweg/ui/src/components/BFlex/BFlexCol.vue
@@ -1,9 +1,0 @@
-<template>
-  <div class="col"><slot /></div>
-</template>
-
-<style scoped>
-.col {
-  width: 100%;
-}
-</style>